Adds mounting options to synced folder
Vagrant mounts the synced folder between host and guest as vagrant user by default when user/group are not specified. This prevents other users especially the apache process from reading and writing to the folder, which in this machine serves as the document root of Apache web server.

So mounting options are added for vagrant to assigned directory ownership to apache user which will allow apache process to read and write to the synced folder.